+++ src/wrap.js
@@ -1,12 +1,13 @@
+import { copyProperties } from './util'
+
 /**
  * Add methods to an object's prototype.
  */
 export default function makeProto (obj, methods) {
-  const proto = {
-    toJSON: () => obj,
-    ...methods
-  }
+  const proto = copyProperties({
+    toJSON: () => obj
+  }, methods)
   const instance = Object.create(proto)
-  Object.assign(instance, obj)
+  copyProperties(instance, obj)
   return instance
 }